+++ Web/Http/HttpHandleService.php
@@ -10,9 +10,13 @@ use Heptacom\HeptaConnect\Core\Web\Http\Contract\HttpHandleFlowHttpHandlersFacto
 use Heptacom\HeptaConnect\Core\Web\Http\Contract\HttpHandlerStackBuilderFactoryInterface;
 use Heptacom\HeptaConnect\Core\Web\Http\Contract\HttpHandlerStackProcessorInterface;
 use Heptacom\HeptaConnect\Core\Web\Http\Contract\HttpHandleServiceInterface;
+use Heptacom\HeptaConnect\Core\Web\Http\Dump\Contract\ServerRequestCycleDumpCheckerInterface;
+use Heptacom\HeptaConnect\Core\Web\Http\Dump\Contract\ServerRequestCycleDumperInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\Web\Http\Contract\HttpHandleContextInterface;
 use Heptacom\HeptaConnect\Portal\Base\Web\Http\Contract\HttpHandlerStackInterface;
+use Heptacom\HeptaConnect\Portal\Base\Web\Http\HttpHandlerStackIdentifier;
+use Heptacom\HeptaConnect\Portal\Base\Web\Http\ServerRequestCycle;
 use Heptacom\HeptaConnect\Storage\Base\Action\WebHttpHandlerConfiguration\Find\WebHttpHandlerConfigurationFindCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\WebHttpHandlerConfiguration\WebHttpHandlerConfigurationFindActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
@@ -22,6 +26,9 @@ use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 use Ramsey\Uuid\Uuid;
 
+/**
+ * @SuppressWarnings(PHPMD.ExcessiveParameterList)
+ */
 final class HttpHandleService implements HttpHandleServiceInterface
 {
     /**
@@ -42,59 +49,88 @@ final class HttpHandleService implements HttpHandleServiceInterface
         private StorageKeyGeneratorContract $storageKeyGenerator,
         private ResponseFactoryInterface $responseFactory,
         private WebHttpHandlerConfigurationFindActionInterface $httpHandlerConfigurationFindAction,
-        private HttpHandleFlowHttpHandlersFactoryInterface $httpHandleFlowHttpHandlersFactory
+        private HttpHandleFlowHttpHandlersFactoryInterface $httpHandleFlowHttpHandlersFactory,
+        private ServerRequestCycleDumpCheckerInterface $dumpChecker,
+        private ServerRequestCycleDumperInterface $requestResponsePairDumper
     ) {
     }
 
     public function handle(ServerRequestInterface $request, PortalNodeKeyInterface $portalNodeKey): ResponseInterface
     {
-        $portalNodeKey = $portalNodeKey->withoutAlias();
-        $path = $request->getUri()->getPath();
+        $httpHandlerStackIdentifier = new HttpHandlerStackIdentifier(
+            $portalNodeKey->withoutAlias(),
+            $request->getUri()->getPath()
+        );
         $response = $this->responseFactory->createResponse(501);
+        $response = $this->handlePortalNodeRequest($httpHandlerStackIdentifier, $request, $response);
+        $requestCycle = new ServerRequestCycle($request, $response);
+
+        if ($this->dumpChecker->shallDump($httpHandlerStackIdentifier, $requestCycle)) {
+            $this->requestResponsePairDumper->dump($httpHandlerStackIdentifier, $requestCycle);
+        }
+
+        return $response;
+    }
+
+    private function handlePortalNodeRequest(
+        HttpHandlerStackIdentifier $stackIdentifier,
+        ServerRequestInterface $request,
+        ResponseInterface $response
+    ): ResponseInterface {
         // TODO push onto global logging context stack
         $correlationId = Uuid::uuid4()->toString();
 
-        $enabledCheck = $this->httpHandlerConfigurationFindAction->find(
-            new WebHttpHandlerConfigurationFindCriteria($portalNodeKey, $path, 'enabled')
-        );
+        foreach (\array_keys($request->getAttributes()) as $attributeKey) {
+            $attributeKey = (string) $attributeKey;
+
+            if (\str_starts_with($attributeKey, self::REQUEST_ATTRIBUTE_PREFIX)) {
+                $request = $request->withoutAttribute($attributeKey);
+            }
+        }
+
+        $enabledCheck = $this->httpHandlerConfigurationFindAction->find(new WebHttpHandlerConfigurationFindCriteria(
+            $stackIdentifier,
+            'enabled'
+        ));
+
         $enabled = (bool) ($enabledCheck->getValue()['value'] ?? true);
 
         if (!$enabled) {
             $this->logger->warning(LogMessage::WEB_HTTP_HANDLE_DISABLED(), [
                 'code' => 1636845085,
-                'path' => $path,
-                'portalNodeKey' => $portalNodeKey,
+                'path' => $stackIdentifier->getPath(),
+                'portalNodeKey' => $stackIdentifier->getPortalNodeKey(),
                 'request' => $request,
                 'web_http_correlation_id' => $correlationId,
             ]);
 
-            return $response->withHeader('X-HeptaConnect-Correlation-Id', $correlationId)->withStatus(423);
-        }
-
-        $stack = $this->getStack($portalNodeKey, $path);
-
-        if (!$stack instanceof HttpHandlerStackInterface) {
-            $this->logger->critical(LogMessage::WEB_HTTP_HANDLE_NO_HANDLER_FOR_PATH(), [
-                'code' => 1636845086,
-                'path' => $path,
-                'portalNodeKey' => $portalNodeKey,
-                'request' => $request,
-                'web_http_correlation_id' => $correlationId,
-            ]);
+            $response = $response->withStatus(423);
         } else {
-            $response = $this->stackProcessor->processStack($request, $response, $stack, $this->getContext($portalNodeKey));
+            $stack = $this->getStack($stackIdentifier);
+
+            if (!$stack instanceof HttpHandlerStackInterface) {
+                $this->logger->critical(LogMessage::WEB_HTTP_HANDLE_NO_HANDLER_FOR_PATH(), [
+                    'code' => 1636845086,
+                    'path' => $stackIdentifier->getPath(),
+                    'portalNodeKey' => $stackIdentifier->getPortalNodeKey(),
+                    'request' => $request,
+                    'web_http_correlation_id' => $correlationId,
+                ]);
+            } else {
+                $response = $this->stackProcessor->processStack($request, $response, $stack, $this->getContext($stackIdentifier->getPortalNodeKey()));
+            }
         }
 
         return $response->withHeader('X-HeptaConnect-Correlation-Id', $correlationId);
     }
 
-    private function getStack(PortalNodeKeyInterface $portalNodeKey, string $path): ?HttpHandlerStackInterface
+    private function getStack(HttpHandlerStackIdentifier $identifier): ?HttpHandlerStackInterface
     {
-        $cacheKey = $this->storageKeyGenerator->serialize($portalNodeKey) . $path;
+        $cacheKey = $this->storageKeyGenerator->serialize($identifier->getPortalNodeKey()->withoutAlias()) . $identifier->getPath();
 
         if (!\array_key_exists($cacheKey, $this->stackCache)) {
             $builder = $this->stackBuilderFactory
-                ->createHttpHandlerStackBuilder($portalNodeKey, $path)
+                ->createHttpHandlerStackBuilder($identifier)
                 ->pushSource();
 
             if ($builder->isEmpty()) {
@@ -102,7 +138,7 @@ final class HttpHandleService implements HttpHandleServiceInterface
             } else {
                 $builder = $builder->pushDecorators();
 
-                foreach ($this->httpHandleFlowHttpHandlersFactory->createHttpHandlers($portalNodeKey, $path) as $handler) {
+                foreach ($this->httpHandleFlowHttpHandlersFactory->createHttpHandlers($identifier) as $handler) {
                     $builder = $builder->push($handler);
                 }
 
@@ -121,7 +157,7 @@ final class HttpHandleService implements HttpHandleServiceInterface
 
     private function getContext(PortalNodeKeyInterface $portalNodeKey): HttpHandleContextInterface
     {
-        $cacheKey = $this->storageKeyGenerator->serialize($portalNodeKey);
+        $cacheKey = $this->storageKeyGenerator->serialize($portalNodeKey->withoutAlias());
         $this->contextCache[$cacheKey] ??= $this->contextFactory->createContext($portalNodeKey);
 
         return clone $this->contextCache[$cacheKey];